GuessTheFlag Custom Image View
Added a Custom image View FlagImage that will be reused for the Images of flags presented to the User that is guessing which flag correlates to which country. Each FlagImage will be resized, provided a new aspectRatio, frame, clipShape, and given a shadow for each flag being shown such that it made sense to reduce code complexity by creating a custom struct for it.