Added start.txt file for root word list; Added input word validation
The app will search for the startWords file on game start and retrieve a random root word to play with. The user can input words based on the root word, incorrect submissions will result in a pop-up alert detailing mistakes, correct submissions will be added to the used words list to be displayed below the TextField.