Added a EditUserView and Created a NavigationStack
ContentView will create a NavigationStack, presenting a List of all created Users. When a User NavigationLink is selected it redirects to that users EditUserView. In the event the Add User + button is pressed it will immediately create a blank user and redirect the path of the Navigation Stack to that blank users EditUserView.