Created a Job model to link with User
Relationship defined between User and Job. User will have a empty Jobs array that will be able to store a reference to that Users assigned jobs. When a user is deleted so will all the Job it has stored in its jobs array due to the .cascade relationship. Jobs will have a optional reference to the User owner.